requirements.txt
Flask==3.0.0
PyPDF2==3.0.1
pdfplumber==0.10.3
python-dotenv==1.0.0
Werkzeug==3.0.1
Pillow==10.1.0
